requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
gunicorn==22.0.0
python-dotenv==1.0.1
Werkzeug==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
psycopg2-binary==2.9.9  # Add this for PostgreSQL support